+++ open_dubbing/main.py
@@ -33,6 +33,11 @@ from open_dubbing.translation_apertium import TranslationApertium
 from open_dubbing.translation_nllb import TranslationNLLB
 from open_dubbing.video_processing import VideoProcessing
 
+WHISPER_MODEL_NAMES = [
+    "medium",
+    "large-v3",
+]
+
 
 def _init_logging():
     # Create a logger
@@ -135,11 +140,13 @@ def list_supported_languages(_tts, translation, device):  # TODO: Not used
 def main():
     _init_logging()
     """Parses command-line arguments and runs the dubbing process."""
-    parser = argparse.ArgumentParser(description="Run the end-to-end dubbing process.")
+    parser = argparse.ArgumentParser(
+        description="AI dubbing system which uses machine learning models to automatically translate and synchronize audio dialogue into different languages"
+    )
     parser.add_argument(
         "--input_file",
         required=True,
-        help="Path to the input video or audio file.",
+        help="Path to the input video file.",
     )
     parser.add_argument(
         "--output_directory",
@@ -214,12 +221,27 @@ def main():
         "--cpu_threads",
         type=int,
         default=0,
-        help="number of threads used for CPU inference (if not specified uses defaults for each framework)",
+        help="number of threads used for CPU inference (if is not specified uses defaults for each framework)",
     )
     parser.add_argument(
         "--debug",
         action="store_true",
-        help="keep intemediate files and generate specific files for debugging",
+        help="keep intermediate files and generate specific files for debugging",
+    )
+
+    parser.add_argument(
+        "--nllb_model",
+        type=str,
+        default="nllb-200-1.3B",
+        choices=["nllb-200-1.3B", "nllb-200-3.3B"],
+        help="NLLB translation model size. 'nllb-200-3.3B' gives best translation quality",
+    )
+
+    parser.add_argument(
+        "--whisper_model",
+        default="medium",
+        choices=WHISPER_MODEL_NAMES,
+        help="name of the OpenAI Whisper speech to text model size to use",
     )
 
     args = parser.parse_args()
@@ -251,13 +273,29 @@ def main():
 
     if args.stt == "auto":
         if sys.platform == "darwin":
-            stt = SpeechToTextWhisperTransfomers(args.device, args.cpu_threads)
+            stt = SpeechToTextWhisperTransfomers(
+                model_name=args.whisper_model,
+                device=args.device,
+                cpu_threads=args.cpu_threads,
+            )
         else:
-            stt = SpeechToTextFasterWhisper(args.device, args.cpu_threads)
+            stt = SpeechToTextFasterWhisper(
+                model_name=args.whisper_model,
+                device=args.device,
+                cpu_threads=args.cpu_threads,
+            )
     elif args.stt == "faster-whisper":
-        stt = SpeechToTextFasterWhisper(args.device, args.cpu_threads)
+        stt = SpeechToTextFasterWhisper(
+            model_name=args.whisper_model,
+            device=args.device,
+            cpu_threads=args.cpu_threads,
+        )
     else:
-        stt = SpeechToTextWhisperTransfomers(args.device, args.cpu_threads)
+        stt = SpeechToTextWhisperTransfomers(
+            model_name=args.whisper_model,
+            device=args.device,
+            cpu_threads=args.cpu_threads,
+        )
 
     stt.load_model()
     source_language = args.source_language
@@ -267,6 +305,7 @@ def main():
 
     if args.translator == "nllb":
         translation = TranslationNLLB(args.device)
+        translation.load_model(args.nllb_model)
     elif args.translator == "apertium":
         server = args.apertium_server
         if len(server) == 0:
@@ -279,7 +318,6 @@ def main():
     else:
         raise ValueError(f"Invalid translator value {args.translator}")
 
-    translation.load_model()
     check_languages(source_language, args.target_language, tts, translation, stt)
 
     if not os.path.exists(args.output_directory):